Added .ssh/config
By breaking the .ssh/config file up into files in .ssh/config.d and
having .ssh/config include them, I can commit universally used ssh
configuration to this repo while still including machine-specific
configuration.

Currently, .ssh/config.d contains a file that has ssh automatically add
SSH keys to the SSH Agent, and a file that forces gitlab.com to use
IPv4, even if IPv6 is available (due to issues encountered in the past
with IPv6 connections to GitLab)

Also, modified .bash_profile to check for an SSH Agent, and start one if
one is not found. Starting the SSH Agent is handled by a desktop
environment on most of my machines, but this may not always be the case
on systems where I'd like to use these dotfiles.